Fix, as per http://david-burger.blogspot.com/2008/10/attachmentfu-with-optional.html

+++ lib/technoweenie/attachment_fu/backends/file_system_backend.rb
@@ -85,9 +85,11 @@ module Technoweenie # :nodoc:
         protected
           # Destroys the file.  Called in the after_destroy callback
           def destroy_file
-            FileUtils.rm full_filename
-            # remove directory also if it is now empty
-            Dir.rmdir(File.dirname(full_filename)) if (Dir.entries(File.dirname(full_filename))-['.','..']).empty?
+            if filename?
+              FileUtils.rm full_filename
+              # remove directory also if it is now empty
+              Dir.rmdir(File.dirname(full_filename)) if (Dir.entries(File.dirname(full_filename))-['.','..']).empty?
+            end
           rescue
             logger.info "Exception destroying  #{full_filename.inspect}: [#{$!.class.name}] #{$1.to_s}"
             logger.warn $!.backtrace.collect { |b| " > #{b}" }.join("\n")
